Feature/encode maps (#1)
* Use struct directly if type matches

* Return errors when decoding negative numbers to unsigned integers.

* Create .travis.yml

* decode hooks can take types or kinds

* simplify code a bit

* string to time duration hook

* bigger test for decode hook with types

* comment

* add test for decoding direct struct

* implement errwrap.Wrapper

* introduce ZeroFields, starting to work with maps

This MIGHT introduce backwards incompatibities but we never meant for it
to work this way to begin with. Maps are now merged by default but you
can configure to zero them.

* add another test

* Fields don't need to be anonymous to support "squash"

* clarify comment

* Add decodeViaJSON, benchmark and example (as a test).

decodeViaJSON performs faster with less memory allocations for simple
uses, despite performing the task using an indirect approach. Use it as
a baseline for comparing performance of mapstructure.Decode.

Credit to @sqs and @gbbr for initial idea.

* Update mapstructure_examples_test.go

* failing test for slice to map

* decode slice of maps to a map

* squash: validate type of sqashed fields also (bug-fix)

If the user adds a 'squash' tag to a non-struct type, the call to
structType.NumField() will silently fail, the application stops.

Fixed by adding an error if the squashed field's type is not struct.

* returning untyped nil from decode hook will assign properly to interface

* Handle json.Number values.

In order to prevent the JSON decoder from converting all values to
`float64`, which causes a number of problems, one must set `UseNumber` on
a JSON decoder which tells it to unmarshal as `json.Number` values.
These can then be converted to ints, floats, or used as strings.

However, although setting this value is very common, mapstructure cannot
currently decode from a `map[string]interface{}` containing a
`json.Number` to a struct with those values represented as int or float.

This adds code and test cases to correctly decode a json.Number into an
int or float as expected. It does not decode into uint, although this
could be handled, since json.Number does not natively decode into an
unsigned int.

* remove redundant variable of for loop

* Embedded non-structs are supported (such as slices)

* Fix typo in documentation

* Tests to reproduce

* #48: fixes for ptr and slice cases

* support func decoding via hooks

* verbose decode hook errors

* fix decode slice in zero fields false case

* Promote single values to slice values for weakly typed input

This automatically promotes single standalone values to slices/arrays
if the target type is a slice or array. This only happens in weakly
typed mode.

The single value is weakly decoded into the target slice type, meaning
that a string value can decode to an int slice for example.

This enables the following to decode:

```
source := map[string]interface{}{"key": "4"}

var target struct {
  Key []int
}

// Decode source to target

// Pseudocode
result == { Key: []int{4} }
```

* Add the ability to decode structs to maps